Split out Core.hs into Processing.hs
Now the user facing API is just Language.LSP.Server
